+++ CHANGELOG.md
@@ -5,6 +5,12 @@
 ### New
 
 - Add unschedule button to Post UI
+- Profile editing screen
+- Allow user to set preferred timezone
+
+### Changes
+
+- Use user's preferred timezone for schedule date picker
 
 ## Version 3.1.1